requirements.txt
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
blinker==1.8.2
cachetools==5.3.3
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
distro==1.9.0
exceptiongroup==1.2.1
Flask==3.0.3
google-api-core==2.19.1
google-auth==2.32.0
google-cloud-core==2.4.1
google-cloud-translate==3.15.4
googleapis-common-protos==1.63.2
grpcio==1.64.1
grpcio-status==1.64.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
importlib_metadata==8.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
MarkupSafe==2.1.5
nltk==3.8.1
openai==1.35.13
proto-plus==1.24.0
protobuf==5.27.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.8.2
pydantic_core==2.20.1
redis==5.0.7
regex==2024.5.15
requests==2.32.3
rsa==4.9
sniffio==1.3.1
tqdm==4.66.4
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.3
zipp==3.19.2
